Python_version/requirements.txt
PyQt6>=6.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
nltk>=3.8.1 